+++ CHANGES.md
@@ -33,6 +33,7 @@
 - Testing: Add `datetime` and `list` updater utils
 - Testing: Add `setLocalizedDescription` action for shipping_method
 - Testing: Add `changeAttributeConstraint` action for product_type
+- Testing: Add `setLineItemCustomField` action for cart and order
 
   
 ## Notes on code generation